--- api/.ebextensions
@@ -1,4 +1,0 @@
-option_settings:
-  - namespace: aws:elasticbeanstalk:container:nodejs
-    option_name: NodeCommand
-    value: "npm start"
