--- api/app.js
@@ -1,60 +1,0 @@
-const createError = require('http-errors');
-const express = require('express');
-const mongoose = require('mongoose');
-const path = require('path');
-const cookieParser = require('cookie-parser');
-const logger = require('morgan');
-const cors = require("cors");
-
-const dotenv = require('dotenv')
-dotenv.config();
-
-const indexRouter = require('./routes/index');
-const usersRouter = require('./routes/users');
-const helloWorldAPI = require('./routes/helloWorldAPI');
-
-const app = express();
-
-// view engine setup
-app.set('views', path.join(__dirname, 'views'));
-app.set('view engine', 'jade');
-
-app.use(cors());
-app.use(logger('dev'));
-app.use(express.json());
-app.use(express.urlencoded({ extended: false }));
-app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'public')));
-
-app.use('/', indexRouter);
-app.use('/users', usersRouter);
-app.use('/helloworld', helloWorldAPI);
-
-const port = process.env.PORT || 8081;
-
-mongoose.connect("mongodb+srv://meanxael:tt5FXQD1XudeUPLd@cluster0.u46go.mongodb.net/group3?retryWrites=true&w=majority", {
-  useUnifiedTopology: true,
-  useNewUrlParser: true,
-}).then(() => {
-  app.listen(port, function () {
-      console.log('Server running at http://127.0.0.1:' + port + '/');
-  });
-})
-
-// catch 404 and forward to error handler
-app.use(function(req, res, next) {
-  next(createError(404));
-});
-
-// error handler
-app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error');
-});
-
-module.exports = app;
