--- api/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "api",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "start": "node app.js"
-  },
-  "engines": { "node" : "14.16.0" },
-  "dependencies": {
-    "concurrently": "^6.4.0",
-    "cookie-parser": "~1.4.4",
-    "cors": "^2.8.5",
-    "debug": "~2.6.9",
-    "dotenv": "^10.0.0",
-    "express": "~4.16.1",
-    "http-errors": "~1.6.3",
-    "jade": "~1.11.0",
-    "mongoose": "^6.1.1",
-    "morgan": "~1.9.1"
-  }
-}
