--- api/views/index.jade
@@ -1,6 +1,0 @@
-extends layout
-
-block content
-  h1= title
-  p Welcome to #{title}
-  p= text
