+++ package.json
@@ -1,0 +1,18 @@
+{
+  "dependencies": {
+    "concurrently": "^6.4.0",
+    "cors": "^2.8.5",
+    "dotenv": "^10.0.0",
+    "express": "^4.17.1",
+    "mongoose": "^6.1.1",
+    "nodemon": "^2.0.15"
+  },
+  "scripts": {
+    "build": "cd group3-app && npm run build",
+    "install-app": "cd group3-app && npm install",
+    "heroku-postbuild": "npm run install-client && npm run build",
+    "server": "nodemon server.js",
+    "develop": "concurrently --kill-others-on-fail \"npm run server\" \"npm run start --prefix group3-app\"",
+    "start": "concurrently --kill-others-on-fail \"npm run server\" \"npm run start --prefix group3-app\""
+  }
+}
